Fix cloze bug & refactor
Fixed a bug with the way the cloze alt shortcut was implemented and separated code into 2 files to reduce clode clutter
