+++ custom_shortcuts/__init__.py
@@ -1,1 +1,2 @@
 from . import custom_shortcuts
+from . import cs_functions
