+++ custom_shortcuts/custom_shortcuts.py
@@ -9,6 +9,7 @@ from aqt.reviewer import Reviewer
 from anki.utils import json
 from bs4 import BeautifulSoup
 import warnings
+from . import cs_functions as functions
 
 #Gets config.json as config
 config = mw.addonManager.getConfig(__name__)
@@ -148,42 +149,6 @@ def cs_mt_setupShortcuts():
     m.actionCreateFiltered.setShortcuts(scuts_list["m_toolbox create filtered deck"])
     m.actionAdd_ons.setShortcuts(scuts_list["m_toolbox addons"])
 
-#Converts json shortcuts into functions for the reviewer
-#sToF: shortcutToFunction
-def review_sToF(self,scut):
-
-    #"reviewer" is retained for copy-pastability, may be removed later
-    # "self.mw.onEditCurrent" is exactly how it was in reviewer.py, DO NOT CHANGE
-    sdict = {
-        "reviewer edit current": self.mw.onEditCurrent,
-        "reviewer flip card": self.onEnterKey,
-        "reviewer flip card 1": self.onEnterKey,
-        "reviewer flip card 2": self.onEnterKey,
-        "reviewer flip card 3": self.onEnterKey,
-        "reviewer options menu": self.onOptions,
-        "reviewer record voice": self.onRecordVoice,
-        "reviewer play recorded voice": self.onReplayRecorded,
-        "reviewer play recorded voice 1": self.onReplayRecorded,
-        "reviewer play recorded voice 2": self.onReplayRecorded,
-        "reviewer delete note": self.onDelete,
-        "reviewer suspend card": self.onSuspendCard,
-        "reviewer suspend note": self.onSuspend,
-        "reviewer bury card": self.onBuryCard,
-        "reviewer bury note": self.onBuryNote,
-        "reviewer mark card": self.onMark,
-        "reviewer set flag 1": lambda: self.setFlag(1),
-        "reviewer set flag 2": lambda: self.setFlag(2),
-        "reviewer set flag 3": lambda: self.setFlag(3),
-        "reviewer set flag 4": lambda: self.setFlag(4),
-        "reviewer set flag 0": lambda: self.setFlag(0),
-        "reviewer replay audio": self.replayAudio,
-        "reviewer choice 1": lambda: self._answerCard(1),
-        "reviewer choice 2": lambda: self._answerCard(2),
-        "reviewer choice 3": lambda: self._answerCard(3),
-        "reviewer choice 4": lambda: self._answerCard(4),
-    }
-    return sdict[scut]
-
 #Governs the shortcuts on the review window
 def cs_review_setupShortcuts(self):
     dupes = []
@@ -233,7 +198,7 @@ def cs_editor_setupShortcuts(self):
         (config_scuts["editor foreground"], self.onForeground),
         (config_scuts["editor change col"], self.onChangeCol),
         (config_scuts["editor cloze"], self.onCloze),
-        (config_scuts["editor cloze alt"], self.onCloze),
+        (config_scuts["editor cloze alt"], self.onAltCloze),
         (config_scuts["editor add media"], self.onAddMedia),
         (config_scuts["editor record sound"], self.onRecSound),
         (config_scuts["editor insert latex"], self.insertLatex),
@@ -343,28 +308,15 @@ def cs_conflictDetect():
         showWarning(conflictStr)
 
 
-#Mimics the style of other Anki functions, analogue of customPaste
-#Note that the saveNow function used earler takes the cursor to the end of the line,
-#as it is meant to save work before entering a new window
-def cs_editor_custom_paste(self):
-    self._customPaste()
-
-#Mimics the style of other Anki functions, analogue of _customPaste
-def cs_uEditor_custom_paste(self):
-    html = config_scuts["Ω custom paste text"]
-    if config_scuts["Ω custom paste end style"].upper() == "Y":
-        html += "</span>\u200b"
-    with warnings.catch_warnings() as w:
-        warnings.simplefilter('ignore', UserWarning)
-        html = str(BeautifulSoup(html, "html.parser"))
-    self.doPaste(html,True,True)
 
 #Functions that execute on startup
+Editor.customPaste = functions.cs_editor_custom_paste
+Editor._customPaste = functions.cs_uEditor_custom_paste
+Editor.onAltCloze = functions.cs_editor_onAltCloze
+Editor._onAltCloze = functions.cs_uEditor_onAltCloze
+Reviewer.sToF = functions.review_sToF
 Editor.setupShortcuts = cs_editor_setupShortcuts
-Editor.customPaste = cs_editor_custom_paste
-Editor._customPaste = cs_uEditor_custom_paste
 Reviewer._shortcutKeys = cs_review_setupShortcuts
-Reviewer.sToF = review_sToF
 
 mw.applyShortcuts = cs_main_setupShortcuts
